added a main process for testing
